+++ modules/jellyfin/default.nix
@@ -8,8 +8,9 @@ with lib;
 let
   inherit (config) nixflix;
   inherit (config.nixflix) globals;
+  inherit (import ../../lib/mkVirtualHosts.nix { inherit lib config; }) mkVirtualHost;
   cfg = config.nixflix.jellyfin;
-  hostname = "${cfg.subdomain}.${nixflix.nginx.domain}";
+  hostname = "${cfg.subdomain}.${nixflix.reverseProxy.domain}";
 
   xml = import ./xml.nix { inherit lib; };
 
@@ -53,8 +54,15 @@ in
     ./usersConfigService.nix
   ];
 
-  config = mkMerge [
-    (mkIf (nixflix.enable && cfg.enable) {
+  config = mkIf (nixflix.enable && cfg.enable) (mkMerge [
+    (mkVirtualHost {
+      inherit hostname;
+      inherit (cfg.reverseProxy) expose;
+      port = cfg.network.internalHttpPort;
+      disableBuffering = true;
+      websocketUpgrade = true;
+    })
+    {
       warnings = pluginResolution.resolutionWarnings;
 
       nixflix.jellyfin = {
@@ -351,40 +359,8 @@ in
         ];
       };
 
-      networking.hosts = mkIf (nixflix.nginx.enable && nixflix.nginx.addHostsEntries) {
-        "127.0.0.1" = [ hostname ];
-      };
-
-      services.nginx.virtualHosts."${hostname}" = mkIf nixflix.nginx.enable {
-        inherit (config.nixflix.nginx) forceSSL;
-        useACMEHost = if config.nixflix.nginx.enableACME then config.nixflix.nginx.domain else null;
-
-        locations =
-          let
-            proxyHost = if cfg.vpn.enable then config.vpnNamespaces.wg.namespaceAddress else "127.0.0.1";
-          in
-          {
-            "/" = {
-              proxyPass = "http://${proxyHost}:${toString cfg.network.internalHttpPort}";
-              recommendedProxySettings = true;
-              extraConfig = ''
-                proxy_set_header X-Real-IP $remote_addr;
-
-                proxy_buffering off;
-              '';
-            };
-            "/socket" = {
-              proxyPass = "http://${proxyHost}:${toString cfg.network.internalHttpPort}";
-              proxyWebsockets = true;
-              recommendedProxySettings = true;
-              extraConfig = ''
-                proxy_set_header X-Real-IP $remote_addr;
-              '';
-            };
-          };
-      };
-    })
-    (mkIf (nixflix.enable && cfg.enable && config.nixflix.vpn.enable && cfg.vpn.enable) {
+    }
+    (mkIf (config.nixflix.vpn.enable && cfg.vpn.enable) {
       systemd.services.jellyfin.vpnConfinement = {
         enable = true;
         vpnNamespace = "wg";
@@ -397,5 +373,5 @@ in
         }
       ];
     })
-  ];
+  ]);
 }